feat: 실시간 쿠폰 sse 기능 추가

+++ src/main/java/kr/hhplus/be/infrastructure/kafka/consumer/CouponDecrementStockConsumer.java
@@ -8,45 +8,59 @@ import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
+import kr.hhplus.be.application.usercoupon.dto.UserCouponResult;
 import kr.hhplus.be.domain.common.event.CouponIssuedEvent;
 import kr.hhplus.be.domain.common.event.CouponIssuedSuccess;
-import kr.hhplus.be.domain.common.event.UserCouponSyncTask;
 import kr.hhplus.be.domain.coupon.repository.CouponRedisRepository;
-import kr.hhplus.be.domain.usercoupon.model.UserCoupon;
 import kr.hhplus.be.domain.usercoupon.repository.UserCouponRedisRepository;
-import kr.hhplus.be.domain.usercoupon.repository.UserCouponRepository;
+import kr.hhplus.be.infrastructure.sse.SseEmitterService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class CouponIssuedConsumer {
+public class CouponDecrementStockConsumer {
 
-	private final UserCouponRepository userCouponRepository;
 	private final CouponRedisRepository couponRedisRepository;
 	private final UserCouponRedisRepository userCouponRedisRepository;
 	private final KafkaTemplate<String, Object> kafkaTemplate;
+	private final SseEmitterService sseEmitterService;
 
 	@KafkaListener(topics = TOPIC_COUPON_ISSUED, groupId = "coupon-issuer-group")
 	public void listen(CouponIssuedEvent event) {
 		Long couponId = event.couponId();
+		String taskId = event.taskId();
 
 		Long decrementStock = couponRedisRepository.decrementStock(couponId);
 
 		if (decrementStock != null && decrementStock >= 0) {
-			log.info("재고 차감 성공. taskId: {}", event.taskId());
+			log.info("재고 차감 성공. taskId: {}", taskId);
 
 			CouponIssuedSuccess successDto = new CouponIssuedSuccess(
-				event.taskId(), event.userId(), event.couponId(), LocalDateTime.now()
+				taskId, event.userId(), event.couponId(), LocalDateTime.now()
 			);
 			kafkaTemplate.send(TOPIC_COUPON_ISSUED_SUCCESS, successDto);
-		} else {
-			if (decrementStock != null) {
-				couponRedisRepository.incrementStock(event.couponId(), 1L);
+
+			if (decrementStock == 0) {
+				log.info("쿠폰 재고 소진 완료: couponId {}", couponId);
+				couponRedisRepository.removeValidCoupon(couponId);
 			}
-			couponRedisRepository.removeValidCoupon(couponId);
-			log.warn("재고 없음. taskId: {}", event.taskId());
+
+		} else {
+			log.warn("재고 없음. taskId: {}", taskId);
+
+			if (decrementStock != null) {
+				couponRedisRepository.incrementStock(couponId, 1L);
+			}
+
+			UserCouponResult.UserCouponConfirmation failDto = UserCouponResult.UserCouponConfirmation.of(
+				"FAIL",
+				"재고 없음",
+				event.issuedAt()
+			);
+			userCouponRedisRepository.saveIssueCouponTaskStatus(taskId, "FAIL", "재고 없음", event.issuedAt());
+			sseEmitterService.sendToUser(taskId, "coupon-status", failDto);
 		}
 	}
 }
